# cloudnet-rest-module/src/main/resources/documentation/swagger.yaml
openapi: 3.1.0
info:
  description: >
    CloudNet's rest API version 3 is the next generation in CloudNet's rest
    history.
    
    
    In order to access rest API v3 you will have to authenticate yourself using basic auth.
    To create a basic auth user use the `rest user create <id> <password>` command on your cloudnet node.
    After authenticating with basic auth on `/auth` you will get a bearer jwt access token that must be used for further
    requests to the api. For further information about the response take a look at the `/auth` endpoint.
    
    MENTION CONFIG HERE
    MENTION CASE-SENS HERE
  version: '3'
  title: CloudNet
  contact:
    name: CloudNet
    url: https://cloudnetservice.eu
  license:
    name: Apache license, version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: '{protocol}://{host}:{port}/api/v3/'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
      host:
        default: 127.0.0.1
      port:
        default: '2812'
security:
  - bearerAuth: [ ]
...
paths:
...
  /serviceVersion/environment:
    get:
      tags:
        - Service Versions
      summary: Lists all available service environments
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_version_read`

          - `cloudnet_rest:service_version_list_environments`
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/ServiceEnvironmentType'
          description: Successful listing of available service environments
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      tags:
        - Service Versions
      summary: Adds a new service environment type to the service type provider
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_version_write`

          - `cloudnet_rest:service_version_environment`
      requestBody:
        required: true
        description: The service environment type to add to the service type provider
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceEnvironmentType'
      responses:
        '204':
          description: The service environment type was registered.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /serviceVersion/{version}:
    parameters:
      - name: version
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Service Versions
      summary: Get a specific service version type
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_version_read`

          - `cloudnet_rest:service_version_get`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceVersionType'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /serviceVersion/load:
    post:
      parameters:
        - name: url
          in: query
          required: false
          description: >
            The url to load the service types from. If no url is provided the
            service types

            loaded are the fallback types compiled into the jar file.
          schema:
            type: string
      tags:
        - Service Versions
      summary: Loads the service types into the service type provider
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:service_version_write`

          - `cloudnet_rest:service_version_load`
      responses:
        '204':
          description: Loaded the service versions successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /serviceVersion/install:
    post:
      parameters:
        - name: force
          description: Whether the install should be forced
          in: query
          schema:
            type: boolean
        - name: cache
          description: Whether the download files should be cached
          in: query
          schema:
            type: boolean
      tags:
        - Service Versions
      summary: Installs a service version to a static service or a service template
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - oneOf:
                    - properties:
                        staticService:
                          type: string
                          example: Lobby-1
                    - properties:
                        template:
                          $ref: '#/components/schemas/ServiceTemplate'
                - oneOf:
                    - properties:
                        serviceVersionType:
                          type: string
                          example: paper
                    - properties:
                        serviceVersionType:
                          $ref: '#/components/schemas/ServiceVersionType'
                - oneOf:
                    - properties:
                        serviceVersion:
                          type: string
                          example:
                    - properties:
                        serviceVersion:
                          $ref: '#/components/schemas/ServiceVersion'
      responses:
        '204':
          description: Installed the service version to the static service / template.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    Problem:
      description: The action couldn't be exececuted
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      description: >-
        An object describing a problem as specific in RFC 9457. Note that this
        object can contain any number of additional properties depending on the
        called route and error cause.
      type: object
      properties:
        type:
          type: string
          format: uri
          description: An absolute URI that identifies the problem type
          default: about:blank
        title:
          type: string
          description: >-
            A short summary of the problem type. Written in English and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized).
          example: Service Unavailable
        status:
          type: integer
          format: int32
          description: >-
            The HTTP status code generated by the origin server for this
            occurrence of the problem.
          minimum: 400
          exclusiveMaximum: 600
          example: 503
        detail:
          type: string
          description: >-
            A human-readable explanation specific to this occurrence of the
            problem
        instance:
          type: string
          format: uri
          description: >-
            An absolute URI reference that identifies the specific occurrence of
            the problem. It may or may not yield further information if
            dereferenced.
...
    JsonDocPropertyable:
      type: object
      properties:
        properties:
          type: object
          examples:
            - { }
...
    ServiceTemplate:
      type: object
      properties:
        prefix:
          type: string
          examples:
            - Global
        name:
          type: string
          examples:
            - server
        storage:
          type: string
          examples:
            - local
        priority:
          type: integer
          examples:
            - 5
        alwaysCopyToStaticServices:
          type: boolean
          examples:
            - false
...
    ServiceVersion:
      type: object
      properties:
        name:
          type: string
          example: latest
        url:
          type: string
          example: >-
            https://api.papermc.io/v2/projects/paper/versions/1.14.4/builds/243/downloads/paper-1.14.4-243.jar
        minJavaVersion:
          type: integer
          example: 8
        maxJavaVersion:
          type: integer
          example: 11
        deprecated:
          type: boolean
          example: false
        cacheFiles:
          type: boolean
          example: false
        properties:
          type: object
          example: { }
        additionalDownloads:
          type: object
          additionalProperties:
            type: string
    ServiceVersionType:
      type: object
      properties:
        name:
          type: string
          example: paperspigot
        environmentType:
          type: string
          example: MINECRAFT_SERVER
        installSteps:
          type: array
          items:
            type: string
          example:
            - DOWNLOAD
            - BUILD
            - UNZIP
            - COPY_FILTER
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ServiceVersion'
    ServiceEnvironmentType:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            name:
              type: string
              example: MINECRAFT_SERVER
            defaultServiceStartPort:
              type: integer
              example: 25565
            defaultProcessArguments:
              type: array
              items:
                type: string
...